Moved some find_pulsar_in_obs.py functions into beam_calc and catalogue_utils
