+++ scripts/find_pulsar_in_obs.py
@@ -19,30 +19,21 @@ which will output a text file called J0534+2200_analytic_beam.txt
 __author__ = 'Nicholas Swainston'
 __date__ = '2016-03-21'
 
-#TODO: (BWM) might be worth trying to only load the necessary parts of modules rather than loading an ENTIRE module. The speed up will be minimal overall, but it's also just a bit neater.
-# You can also just load module/part of modules internally within functions. So if a module is only used once, just load it in the local function rather than up here (globally).
-
-#comon
 import os
 import sys
-import math
 import argparse
 import numpy as np
 import csv
 
-#astropy
-from astropy.coordinates import SkyCoord
-from astropy import units as u
-
-#MWA scripts
 from vcstools import data_load
-from vcstools.pointing_utils import sex2deg, deg2sex, format_ra_dec
-
+from vcstools.pointing_utils import sex2deg, format_ra_dec
 import vcstools.sn_flux_est as sfe
-from mwa_pb import primary_beam
-from vcstools.metadb_utils import mwa_alt_az_za, get_common_obs_metadata,\
-                             get_obs_array_phase, find_obsids_meta_pages,\
-                             getmeta
+from vcstools.metadb_utils import get_obs_array_phase, singles_source_search,
+                                  find_obsids_meta_pages
+from vcstools.catalogue_utils import get_psrcat_ra_dec, grab_source_alog,
+                                     get_rFRB_info
+from vcstools.beam_calc import beam_enter_exit, get_beam_power_over_time,
+                               find_sources_in_obs
 
 
 import logging
@@ -65,324 +56,6 @@ def yes_no(answer):
            return False
         else:
            logger.warning("Please respond with 'yes' or 'no'\n")
-
-
-def get_psrcat_ra_dec(pulsar_list=None, max_dm=250., include_dm=False, query=None):
-    """
-    Uses PSRCAT to return a list of pulsar names, ras and decs. Not corrected for proper motion.
-    Removes pulsars without any RA or DEC recorded
-    If no pulsar_list given then returns all pulsar on the catalogue
-    If include_dm is True then also ouput DM
-
-    get_psrcat_ra_dec(pulsar_list = None)
-    Args:
-        pulsar_list: A space list of pulsar names eg: [J0534+2200, J0538+2817].
-               (default: uses all pulsars)
-    return [[Jname, RAJ, DecJ]]
-    """
-    import psrqpy
-
-    #params = ['JNAME', 'RAJ', 'DECJ', 'DM']
-    if query is None:
-        query = psrqpy.QueryATNF(params = ['PSRJ', 'RAJ', 'DECJ', 'DM'], psrs=pulsar_list, loadfromdb=data_load.ATNF_LOC).pandas
-
-    pulsar_ra_dec = []
-    for i, _ in enumerate(query["PSRJ"]):
-        # Only record if under the max_dm
-        dm = query["DM"][i]
-        if not math.isnan(dm):
-            if float(dm) < max_dm:
-                if include_dm:
-                    pulsar_ra_dec.append([query["PSRJ"][i], query["RAJ"][i], query["DECJ"][i], dm])
-                else:
-                    pulsar_ra_dec.append([query["PSRJ"][i], query["RAJ"][i], query["DECJ"][i]])
-
-
-    return pulsar_ra_dec
-
-
-def grab_source_alog(source_type='Pulsar', pulsar_list=None, max_dm=1000., include_dm=False, query=None):
-    """
-    Will search different source catalogues and extract all the source names, RAs, Decs and, if requested, DMs
-
-    Parameters
-    ----------
-    source_type: string
-        The type of source you would like to get the catalogue for.
-        Your choices are: ['Pulsar', 'FRB', 'rFRB', 'POI' 'RRATs', 'Fermi']
-        Default: 'Pulsar'
-    pulsar_list: list
-        List of sources you would like to extract data for.
-        If None is given then it will search for all available sources
-        Default: None
-    max_dm: float
-        If the source_type is 'Pulsar' then you can set a maximum dm and the function will only
-        return pulsars under that value.
-        Default: 1000.
-    include_dm: Bool
-        If True the function will also return the DM if it is available at the end of the list
-        Default: False
-
-    Returns
-    -------
-    result: list list
-        A list for each source which contains a [source_name, RA, Dec (, DM)]
-        where RA and Dec are in the format HH:MM:SS
-    """
-    modes = ['Pulsar', 'FRB', 'rFRB', 'POI', 'RRATs', 'Fermi']
-    if source_type not in modes:
-        logger.error("Input source type not in known catalogues types. Please choose from: {0}".format(modes))
-        return None
-
-    #Get each source type into the format [[name, ra, dec]]
-    name_ra_dec = []
-    if source_type == 'Pulsar':
-        name_ra_dec = get_psrcat_ra_dec(pulsar_list=pulsar_list, max_dm=max_dm, include_dm=include_dm, query=query)
-
-    elif source_type == 'FRB':
-        import json
-        import urllib.request
-        try:
-            frb_data = json.load(urllib.request.urlopen('http://frbcat.org/products?search=&min=0&max=1000&page=1'))
-        except urllib.error.HTTPError:
-            logger.error('http://frbcat.org/ not available. Returning empty list')
-            # putting and FRB at 90 dec which we should never be able to detect
-            name_ra_dec = [['fake', "00:00:00.00", "90:00:00.00", 0.0]]
-        else:
-            for frb in frb_data['products']:
-                name = frb['frb_name']
-                logger.debug('FRB name: {}'.format(name))
-                ra   = frb['rop_raj']
-                dec  = frb['rop_decj']
-                dm   = frb['rmp_dm'].split("&")[0]
-                if include_dm:
-                    name_ra_dec.append([name, ra, dec, dm])
-                else:
-                    name_ra_dec.append([name, ra, dec])
-
-    elif source_type == "rFRB":
-        info = get_rFRB_info(name=pulsar_list)
-        if info is not None:
-            for line in info:
-                if include_dm:
-                    name_ra_dec.append([line[0], line[1], line[2], line[3]])
-                else:
-                    name_ra_dec.append([line[0], line[1], line[2]])
-
-    elif source_type == "POI":
-        #POI = points of interest
-        db = open(data_load.POI_CSV, "r")
-        for line in db.readlines():
-            if not line.startswith("#"):
-                line = line.split(",")
-                name_ra_dec.append([line[0], line[1], line[2]])
-
-    elif source_type == 'RRATs':
-        import urllib.request
-        try:
-            rrats_data = urllib.request.urlopen('http://astro.phys.wvu.edu/rratalog/rratalog.txt').read().decode()
-        except urllib.error.URLError:
-            logger.error('http://astro.phys.wvu.edu/rratalog/ not available. Returning empty list')
-            # putting and RRAT at 90 dec which we should never be able to detect
-            name_ra_dec = [['fake', "00:00:00.00", "90:00:00.00", 0.0]]
-        else:
-            for rrat in rrats_data.split("\n")[1:-1]:
-                columns = rrat.strip().replace(" ", '\t').split('\t')
-                rrat_cat_line = []
-                if pulsar_list == None or (columns[0] in pulsar_list):
-                    for entry in columns:
-                        if entry not in ['', ' ', '\t']:
-                            rrat_cat_line.append(entry.replace('--',''))
-                    #Removing bad formating for the database
-                    ra = rrat_cat_line[4]
-                    if ra.endswith(":"):
-                        ra = ra[:-1]
-                    dec = rrat_cat_line[5]
-                    if dec.endswith(":"):
-                        dec = dec[:-1]
-
-                    if include_dm:
-                        name_ra_dec.append([rrat_cat_line[0], ra, dec, rrat_cat_line[3]])
-                    else:
-                        name_ra_dec.append([rrat_cat_line[0], ra, dec])
-
-    elif source_type == 'Fermi':
-        # read the fermi targets file
-        try:
-            fermi_loc = os.environ['FERMI_CAND_FILE']
-        except:
-            logger.warning("Fermi candidate file location not found. Returning nothing")
-            return []
-        with open(fermi_loc,"r") as fermi_file:
-            csv_reader = csv.DictReader(fermi_file)
-            for fermi in csv_reader:
-                name = fermi['Source Name'].split()[-1]
-                ra  = fermi[' RA J2000']
-                dec = fermi[' Dec J2000']
-                raj, decj = deg2sex(float(ra), float(dec))
-                pos_u = float(fermi[' a (arcmin)'])
-                if include_dm:
-                    # this actually returns the position uncertainty not dm
-                    name_ra_dec.append([name, raj, decj, pos_u])
-                else:
-                    name_ra_dec.append([name, raj, decj])
-
-    #Remove all unwanted sources
-    if pulsar_list is not None:
-        rrat_cat_filtered = []
-        for line in name_ra_dec:
-            if line[0] in pulsar_list:
-                rrat_cat_filtered.append(line)
-        name_ra_dec = rrat_cat_filtered
-
-    return name_ra_dec
-
-
-def get_rFRB_info(name=None):
-    """
-    Gets repeating FRB info from the csv file we maintain.
-    Input:
-        name: a list of repeating FRB names to get info for. The default is None
-              which gets all rFRBs in the catalogue.
-    Output:
-        [[name, ra, dec, dm, dm error]]
-        A list of all the FRBs which each have a list contining name, ra, dec,
-        dm and dm error
-    """
-    output = []
-    db = open(data_load.KNOWN_RFRB_CSV, "r")
-    for line in db.readlines():
-        if not line.startswith("#"):
-            line = line.split(",")
-            #some FRBs end with a J name. We will ignore these when comparing
-            #by using the first 9 characters
-            FRB = line[0]
-            if name is None:
-                #No input FRBs so return all FRBs
-                output.append(line)
-            elif FRB in name:
-                output.append(line)
-    return output
-
-
-
-def singles_source_search(ra, dec=None, box_size=45.):
-    """
-    Used to find all obsids within a box around the source to make searching through obs_ids more efficient.
-
-    singles_source_search(ra, dec=None, box_size=45.)
-    Parameters:
-    ----------
-    ra: float
-        Right Acension of the source in degrees
-    dec: float
-        Declination of the source in degrees. By default will use the enitre declination range to account for grating lobes
-    box_size: float
-        Radius of the search box. Default: 45
-
-    Returns:
-    --------
-    obsid_metadata: list
-        List of of the metadata for each obsid. The metadata is in the same format as getmeta's output
-    """
-    ra = float(ra)
-    m_o_p = False # moved over (north or south) pole
-
-    if dec is None:
-        dec_top = 90.
-        dec_bot = -90.
-    else:
-        dec = float(dec)
-        dec_top = dec + box_size
-        if dec_top > 90.:
-            dec_top = 90.
-            m_o_p = True
-
-        dec_bot = dec - box_size
-        if dec_top < -90.:
-            dec_top = -90.
-            m_o_p = True
-
-    if m_o_p:
-        obsid_list = find_obsids_meta_pages(params={'mode':'VOLTAGE_START',
-                                                    'minra':0., 'maxra':360.,
-                                                    'mindec':dec_bot,'maxdec':dec_top})
-    else:
-        ra_low = ra - 30. - box_size #30 is the how far an obs would drift in 2 hours(used as a max)
-        ra_high = ra + box_size
-        if ra_low < 0.:
-            ra_new = 360 + ra_low
-            obsid_list = find_obsids_meta_pages(params={'mode':'VOLTAGE_START',
-                                                        'minra':ra_new, 'maxra':360.,
-                                                        'mindec':dec_bot,'maxdec':dec_top})
-            temp_obsid_list = find_obsids_meta_pages(params={'mode':'VOLTAGE_START',
-                                                             'minra':0.,'maxra':ra_high,
-                                                             'mindec':dec_bot,'maxdec':dec_top})
-            for row in temp_obsid_list:
-                obsid_list.append(row)
-        elif ra_high > 360:
-            ra_new = ra_high - 360
-            obsid_list = find_obsids_meta_pages(params={'mode':'VOLTAGE_START',
-                                                        'minra':ra_low, 'maxra':360.,
-                                                        'mindec':dec_bot,'maxdec':dec_top})
-            temp_obsid_list = find_obsids_meta_pages(params={'mode':'VOLTAGE_START',
-                                                             'minra':0., 'maxra':ra_new,
-                                                             'mindec':dec_bot,'maxdec':dec_top})
-            for row in temp_obsid_list:
-                obsid_list.append(row)
-        else:
-            obsid_list = find_obsids_meta_pages(params={'mode':'VOLTAGE_START',
-                                                        'minra':ra_low, 'maxra':ra_high,
-                                                        'mindec':dec_bot,'maxdec':dec_top})
-    return obsid_list
-
-
-def beam_enter_exit(powers, duration, dt=296, min_power=0.3):
-    """
-    Calculates when the source enters and exits the beam
-
-    beam_enter_exit(min_power, powers, imax, dt):
-        powers: list of powers fo the duration every dt and freq powers[times][freqs]
-        dt: the time interval of how often powers are calculated
-        duration: duration of the observation according to the metadata in seconds
-        min_power: zenith normalised power cut off
-    """
-    from scipy.interpolate import UnivariateSpline
-    time_steps = np.array(range(0, duration, dt), dtype=float)
-
-    #For each time step record the min power so even if the source is in
-    #one freq channel it's recorded
-    powers_freq_min = []
-    for p in powers:
-        powers_freq_min.append(float(min(p) - min_power))
-
-    if min(powers_freq_min) > 0.:
-        enter = 0.
-        exit = 1.
-    else:
-        powers_freq_min = np.array(powers_freq_min)
-        logger.debug("time_steps: {}".format(time_steps))
-        logger.debug("powers: {}".format(powers_freq_min))
-        try:
-            spline = UnivariateSpline(time_steps, powers_freq_min , s=0.)
-        except:
-            return None, None
-        if len(spline.roots()) == 2:
-            enter, exit = spline.roots()
-            enter /= duration
-            exit /= duration
-        elif len(spline.roots()) == 1:
-            if powers_freq_min[0] > powers_freq_min[-1]:
-                #power declines so starts in beem then exits
-                enter = 0.
-                exit = spline.roots()[0]/duration
-            else:
-                enter = spline.roots()[0]/duration
-                exit = 1.
-        else:
-            enter = 0.
-            exit = 1.
-    return enter, exit
 
 
 def cal_on_database_check(obsid):
@@ -410,212 +83,6 @@ def cal_on_database_check(obsid):
 
     return check_result
 
-
-def get_beam_power_over_time(beam_meta_data, names_ra_dec,
-                             dt=296, centeronly=True, verbose=False,
-                             option='analytic', degrees=False,
-                             start_time=0):
-    """
-    Calulates the power (gain at coordinate/gain at zenith) for each source over time.
-
-    get_beam_power_over_time(beam_meta_data, names_ra_dec,
-                             dt=296, centeronly=True, verbose=False,
-                             option = 'analytic')
-    Args:
-        beam_meta_data: [obsid,ra, dec, time, delays,centrefreq, channels]
-                        obsid metadata obtained from meta.get_common_obs_metadata
-        names_ra_dec: and array in the format [[source_name, RAJ, DecJ]]
-        dt: time step in seconds for power calculations (default 296)
-        centeronly: only calculates for the centre frequency (default True)
-        verbose: prints extra data to (default False)
-        option: primary beam model [analytic, advanced, full_EE]
-        start_time: the time in seconds from the begining of the observation to
-                    start calculating at
-    """
-    obsid, _, _, time, delays, centrefreq, channels = beam_meta_data
-    names_ra_dec = np.array(names_ra_dec)
-    logger.info("Calculating beam power for OBS ID: {0}".format(obsid))
-
-    starttimes=np.arange(start_time,time+start_time,dt)
-    stoptimes=starttimes+dt
-    stoptimes[stoptimes>time]=time
-    Ntimes=len(starttimes)
-    midtimes=float(obsid)+0.5*(starttimes+stoptimes)
-
-    if not centeronly:
-        PowersX=np.zeros((len(names_ra_dec),
-                             Ntimes,
-                             len(channels)))
-        PowersY=np.zeros((len(names_ra_dec),
-                             Ntimes,
-                             len(channels)))
-        # in Hz
-        frequencies=np.array(channels)*1.28e6
-    else:
-        PowersX=np.zeros((len(names_ra_dec),
-                             Ntimes,1))
-        PowersY=np.zeros((len(names_ra_dec),
-                             Ntimes,1))
-        if centrefreq > 1e6:
-            logger.warning("centrefreq is greater than 1e6, assuming input with units of Hz.")
-            frequencies=np.array([centrefreq])
-        else:
-            frequencies=np.array([centrefreq])*1e6
-    if degrees:
-        RAs = np.array(names_ra_dec[:,1],dtype=float)
-        Decs = np.array(names_ra_dec[:,2],dtype=float)
-    else:
-        RAs, Decs = sex2deg(names_ra_dec[:,1],names_ra_dec[:,2])
-
-    if len(RAs)==0:
-        sys.stderr.write('Must supply >=1 source positions\n')
-        return None
-    if not len(RAs)==len(Decs):
-        sys.stderr.write('Must supply equal numbers of RAs and Decs\n')
-        return None
-    if verbose is False:
-        #Supress print statements of the primary beam model functions
-        sys.stdout = open(os.devnull, 'w')
-    for itime in range(Ntimes):
-        # this differ's from the previous ephem_utils method by 0.1 degrees
-        _, Azs, Zas = mwa_alt_az_za(midtimes[itime], ra=RAs, dec=Decs, degrees=True)
-        # go from altitude to zenith angle
-        theta = np.radians(Zas)
-        phi = np.radians(Azs)
-        for ifreq in range(len(frequencies)):
-            #Decide on beam model
-            if option == 'analytic':
-                rX,rY=primary_beam.MWA_Tile_analytic(theta, phi,
-                                                     freq=frequencies[ifreq], delays=delays,
-                                                     zenithnorm=True,
-                                                     power=True)
-            elif option == 'advanced':
-                rX,rY=primary_beam.MWA_Tile_advanced(theta, phi,
-                                                     freq=frequencies[ifreq], delays=delays,
-                                                     zenithnorm=True,
-                                                     power=True)
-            elif option == 'full_EE':
-                rX,rY=primary_beam.MWA_Tile_full_EE(theta, phi,
-                                                     freq=frequencies[ifreq], delays=delays,
-                                                     zenithnorm=True,
-                                                     power=True)
-        PowersX[:,itime,ifreq]=rX
-        PowersY[:,itime,ifreq]=rY
-    if verbose is False:
-        sys.stdout = sys.__stdout__
-    Powers=0.5*(PowersX+PowersY)
-    return Powers
-
-
-def find_sources_in_obs(obsid_list, names_ra_dec,
-                        obs_for_source=False, dt_input=100, beam='analytic',
-                        min_power=0.3, cal_check=False, all_volt=False,
-                        degrees_check=False, metadata_list=None):
-    """
-    Either creates text files for each MWA obs ID of each source within it or a text
-    file for each source with each MWA obs is that the source is in.
-    Args:
-        obsid_list: list of MWA obs IDs
-        names_ra_dec: [[source_name, ra, dec]]
-        dt: the time step in seconds to do power calculations
-        beam: beam simulation type ['analytic', 'advanced', 'full_EE']
-        min_power: if above the minium power assumes it's in the beam
-        cal_check: checks the MWA pulsar database if there is a calibration for the obsid
-        all_volt: Use all voltages observations including some inital test data
-                  with incorrect formats
-        degrees_check: if false ra and dec is in hms, if true in degrees
-    Output [output_data, obsid_meta]:
-        output_data: The format of output_data is dependant on obs_for_source.
-                     If obs_for_source is True:
-                        output_data = {jname:[[obsid, duration, enter, exit, max_power],
-                                              [obsid, duration, enter, exit, max_power]]}
-                     If obs_for_source is False:
-                        ouput_data = {obsid:[[jname, enter, exit, max_power],
-                                             [jname, enter, exit, max_power]]}
-        obsid_meta: a list of the output of get_common_obs_metadata for each obsid
-    """
-    #prepares metadata calls and calculates power
-    powers = []
-    #powers[obsid][source][time][freq]
-    obsid_meta = []
-    obsid_to_remove = []
-
-    for i, obsid in enumerate(obsid_list):
-        if metadata_list:
-            beam_meta_data, full_meta = metadata_list[i]
-        else:
-            beam_meta_data, full_meta = get_common_obs_metadata(obsid, return_all=True)
-        #beam_meta_data = obsid,ra_obs,dec_obs,time_obs,delays,centrefreq,channels
-
-        if dt_input * 4 >  beam_meta_data[3]:
-            # If the observation time is very short then a smaller dt time is required
-            # to get enough ower imformation
-            dt = int(beam_meta_data[3] / 4.)
-        else:
-            dt = dt_input
-        logger.debug("obsid: {0}, time_obs {1} s, dt {2} s".format(obsid, beam_meta_data[3], dt))
-
-        #check for raw volatge files
-        filedata = getmeta(service='data_files', params={'obs_id':obsid, 'nocache':1})
-        if filedata is None:
-            logger.warning("No file data for obsid {}. Skipping".format(obsid))
-            obsid_to_remove.append(obsid)
-            continue
-        keys = filedata.keys()
-        check = False
-        for k in keys:
-            if '.dat' in k: #TODO check if is still robust
-                check = True
-        if check or all_volt:
-            powers.append(get_beam_power_over_time(beam_meta_data, names_ra_dec,
-                                    dt=dt, centeronly=True, verbose=False,
-                                    option=beam, degrees=degrees_check))
-            obsid_meta.append(beam_meta_data)
-        else:
-            logger.warning('No raw voltage files for %s' % obsid)
-            obsid_to_remove.append(obsid)
-    for otr in obsid_to_remove:
-        obsid_list.remove(otr)
-
-    #chooses whether to list the source in each obs or the obs for each source
-    output_data = {}
-    if obs_for_source:
-        for sn, source in enumerate(names_ra_dec):
-            source_data = []
-            for on, obsid in enumerate(obsid_list):
-                source_ob_power = powers[on][sn]
-                if max(source_ob_power) > min_power:
-                    duration = obsid_meta[on][3]
-                    centre_freq = obsid_meta[on][5] #MHz
-                    channels = obsid_meta[on][6]
-                    bandwidth = (channels[-1] - channels[0] + 1.)*1.28 #MHz
-                    logger.debug("Running beam_enter_exit on obsid: {}".format(obsid))
-                    enter, exit = beam_enter_exit(source_ob_power,duration,
-                                                  dt=dt, min_power=min_power)
-                    if enter is not None:
-                        source_data.append([obsid, duration, enter, exit,
-                                            max(source_ob_power)[0],
-                                            centre_freq, bandwidth])
-            # For each source make a dictionary key that contains a list of
-            # lists of the data for each obsid
-            output_data[source[0]] = source_data
-
-    else:
-        #output a list of sorces for each obs
-        for on, obsid in enumerate(obsid_list):
-            duration = obsid_meta[on][3]
-            obsid_data = []
-            for sn, source in enumerate(names_ra_dec):
-                source_ob_power = powers[on][sn]
-                if max(source_ob_power) > min_power:
-                    enter, exit = beam_enter_exit(source_ob_power, duration,
-                                                  dt=dt, min_power=min_power)
-                    obsid_data.append([source[0], enter, exit, max(source_ob_power)[0]])
-            # For each obsid make a dictionary key that contains a list of
-            # lists of the data for each source/pulsar
-            output_data[obsid] = obsid_data
-
-    return output_data, obsid_meta
 
 def write_output_source_files(output_data,
                               beam='analytic', min_power=0.3, cal_check=False,
