+++ scripts/process_vcs.py
@@ -1,6 +1,5 @@
 #!/usr/bin/env python3
 
-
 import subprocess
 import os
 import sys
