+++ scripts/recombine.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python3
-# vim: tabstop=8 expandtab shiftwidth=4 softtabstop=4
+
 import subprocess
 import os
 import sys
